Move getEventsQuerySkip and getEventsQueryVariables to events utils file
